Support partials with host application routes
It's possible for the host application's sample block to render an
Action View partial that links to a Real Path.

It wouldn't be reasonable to expect the implementation-side view
partials to prefix their routes with `main_app` to differentiate between
themselves and the engine.

This commit declares a `Showcase::RoutingHelper` to expand the set of
routes available to templates rendered by `Showcase::EngineController`
(including `Showcase::PagesController`).

[main_app]: https://guides.rubyonrails.org/engines.html#routes