Call preview_path route through Engine's mounted helper name
Happened on a client that used Rails 6.1.x where Sprockets somehow would intercept the route
and call the assets route. I couldn't figure out how to get to the root of the issue,
but this seems to fix it while working on newer Rails versions.

We may be able to remove this once Rails 7.0+ is the minimum version requirement.

Note: due to how Rails Engine's work people can break our code by
doing `mount Showcase::Engine, as: :another_name` as the mounted helper will
then be `another_name` instead of `showcase`.