Simplify _sample border handling
We can overflow-clip to prevent child element corners from breaking our border at the top and bottom.
This also removes the need for the sc-rounded-t-md on the header.

Additionally it seems like the border on the preview section was convoluted but also didn't have any effect,
so we can trim it.